requirements.txt
easyocr
pytesseract
pdf2image
pillow
python-docx
PyMuPDF
